+++ NEWS.md
@@ -1,3 +1,10 @@
+# *dynaSpec 1.0.3*
+
+* Fix logic for crop and xLim in prep_static_ggspectro()
+* Other fixes to have more predictable behavior of exported MP4 videos using "Matt's approach" with paged_spectro()
+* Added title parameter to add a title to both static and subsequent paged spectrogram (MP4)
+* Added resampleRate parameter to prep_static_ggspectro() for control over resolution of spectrogram (and to speed things up)
+
 # *dynaSpec 1.0.1*
 
 * Fix resampling issue when sampling rate of waves != 44.1 kHz